+++ app/AdminHotel.php
@@ -22,4 +22,5 @@ class AdminHotel extends Model
     'password',
     'remember_token'
     ];
+    
 }
