+++ resources/lang/en/messages.php
@@ -1,16 +1,16 @@
 <?php
-
 return [
 
-    'edit_success'=>'Successfully updated information',
-    'create_success'=>'You had successfully created an account',
+    'edit_success'=>'Updated Successfully',
+    'create_success'=>'Created Successfully',
     'create_admin_hotel'=>'Create a hotel admin',
     'edit_admin_hotel'=>'Edit a hotel admin',
-        'name'=>'Name',
-        'username'=>'Username',
-        'password'=>'Password',
-        'phone'=>'Phone',
-        'email'=>'Email',
-        'hotel'=>'Hotel',
+    'name'=>'Name',
+    'username'=>'Username',
+    'password'=>'Password',
+    'phone'=>'Phone',
+    'email'=>'Email',
+    'hotel'=>'Hotel',
+    'hotel_admin'=>'Hotel admin'
 
 ];
