--- resources/lang/jp/messages.php
@@ -1,7 +1,0 @@
-<?php
-return [
-
-    'edit-success'=>'正常に更新された情報',
-    'create-success'=>'あなたが正常にアカウントを作成していました',
-
-];
